chore(template): remove inlined ZipFile for all workload Lambda fns
