+++ internal/pkg/deploy/cloudformation/stack/testdata/workloads/backend/simple-template.yml
@@ -296,8 +296,7 @@ Resources:
   AddonsStack:
     Metadata:
       "aws:copilot:description": "An Addons CloudFormation Stack for your additional AWS resources"
-    Type: AWS::CloudFormation::Stack # Needed for #1848
-    DependsOn: EnvControllerAction
+    Type: AWS::CloudFormation::Stack
     Condition: HasAddons
     Properties:
       Parameters:
@@ -318,8 +317,8 @@ Resources:
     Type: AWS::Lambda::Function
     Properties:
       Code:
-        ZipFile: |
-          envController
+        S3Bucket:
+        S3Key:
       Handler: "index.handler"
       Timeout: 900
       MemorySize: 512
