+++ internal/pkg/template/templates/workloads/partials/cf/addons.yml
@@ -2,7 +2,7 @@ AddonsStack:
   Metadata:
     'aws:copilot:description': 'An Addons CloudFormation Stack for your additional AWS resources'
   Type: AWS::CloudFormation::Stack
-{{- if ne .EnvControllerLambda ""}} # Needed for #1848
+{{- if gt (len (envControllerParams .)) 0}}
   DependsOn: EnvControllerAction
 {{- end}}
   Condition: HasAddons
