+++ internal/pkg/template/templates/workloads/services/rd-web/cf.yml
@@ -154,9 +154,11 @@ Resources:
   CustomDomainFunction:
     Type: AWS::Lambda::Function
     Properties:
+      {{- with $cr := index .CustomResources "CustomDomainFunction" }}
       Code:
-        S3Bucket: {{ .ScriptBucketName }}
-        S3Key: {{ .CustomDomainLambda }}
+        S3Bucket: {{$cr.Bucket}}
+        S3Key: {{$cr.Key}}
+      {{- end }}
       Handler: "index.handler"
       Timeout: 900
       MemorySize: 512
@@ -213,7 +215,7 @@ Resources:
 {{- end }}
 {{include "publish" . | indent 2}}
 
-{{- if ne .EnvControllerLambda ""}}
+{{- if gt (len (envControllerParams .)) 0}}
 {{include "env-controller" . | indent 2}}
 {{- end}}
